requirements.txt
amqp==5.2.0
asgiref==3.8.1
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.34.114
botocore==1.34.114
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==3.2
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.1
jmespath==1.0.1
kombu==5.3.7
prompt_toolkit==3.0.45
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
